requirements.txt
flask
flask-Bootstrap
flask-redis
flask-session
flask-WTF
inflection
Jinja2
kubernetes
PyYaml
redis
requests
Werkzeug
WTForms
